# kogwistar/server/auth/router.py
from __future__ import annotations

import os
from collections.abc import Mapping
from urllib.parse import urlsplit

from fastapi import APIRouter, Depends, HTTPException, Request
from fastapi.responses import RedirectResponse

from .oidc import OIDCClient
from .service import AuthService
from kogwistar.json_types import JsonValue

router = APIRouter(prefix="/api/auth", tags=["auth"])


def _required_string(claims: Mapping[str, JsonValue], name: str) -> str:
    value = claims.get(name)
    if not isinstance(value, str) or not value:
        raise HTTPException(status_code=401, detail=f"OIDC claim {name!r} is invalid")
    return value


def _optional_string(claims: Mapping[str, JsonValue], name: str) -> str | None:
    value = claims.get(name)
    return value if isinstance(value, str) else None


def get_auth_service(request: Request) -> AuthService:
    return request.app.state.auth_service


def _get_oidc_clients(request: Request) -> dict[str, OIDCClient]:
    return getattr(request.app.state, "oidc_clients", None) or {}


def _get_default_provider(request: Request) -> str | None:
    return getattr(request.app.state, "oidc_default_provider", None)


def _resolve_provider_name(request: Request, provider: str | None = None) -> str:
    clients = _get_oidc_clients(request)
    provider_name = provider or _get_default_provider(request)
    if not provider_name:
        raise HTTPException(status_code=400, detail="OIDC provider is not configured")
    if provider_name not in clients:
        raise HTTPException(
            status_code=400,
            detail=f"OIDC provider {provider_name!r} is not configured",
        )
    return provider_name


def _get_oidc_client_for_provider(
    request: Request, provider: str | None = None
) -> tuple[str, OIDCClient]:
    provider_name = _resolve_provider_name(request, provider)
    oidc = _get_oidc_clients(request).get(provider_name)
    if oidc is None:
        raise HTTPException(
            status_code=400,
            detail=f"OIDC provider {provider_name!r} is not configured",
        )
    return provider_name, oidc


def _get_auth_mode(request: Request) -> str:
    return (getattr(request.app.state, "auth_mode", None) or "oidc").lower()


def _allowed_return_targets() -> tuple[str, ...]:
    raw = os.getenv("AUTH_ALLOWED_RETURN_URLS", "")
    configured = tuple(item.strip() for item in raw.split(",") if item.strip())
    if configured:
        return configured
    return ((os.getenv("UI_URL", "/") or "/").strip(),)


def _resolve_return_target(value: str | None) -> str:
    target = (value or os.getenv("UI_URL", "/") or "/").strip()
    if target.startswith("/") and not target.startswith("//"):
        return target
    parsed = urlsplit(target)
    if not parsed.scheme or not parsed.netloc or parsed.username or parsed.password:
        raise HTTPException(status_code=400, detail="Invalid return_to target")
    if target not in _allowed_return_targets():
        raise HTTPException(status_code=400, detail="return_to target is not allowed")
    return target


def _mint_dev_token(auth_service: AuthService) -> str:
    email = os.getenv("DEV_AUTH_EMAIL", "dev@example.com")
    # Keep the historical seeded identity by default, but make an explicit
    # development email identify a distinct local user unless a subject was
    # deliberately supplied.
    subject = os.getenv("DEV_AUTH_SUBJECT") or (
        email if os.getenv("DEV_AUTH_EMAIL") else "dev"
    )
    display_name = os.getenv("DEV_AUTH_NAME", "Dev User")
    role = os.getenv("DEV_AUTH_ROLE", "ro")
    ns_raw = os.getenv("DEV_AUTH_NS")
    caps_raw = os.getenv("DEV_AUTH_CAPABILITIES")
    if ns_raw:
        ns = ns_raw.split(",") if "," in ns_raw else ns_raw
    else:
        # Default to all namespaces in dev mode to satisfy all apps/tests
        ns = ["docs", "conversation", "workflow", "wisdom"]
    if caps_raw:
        capabilities = caps_raw.split(",") if "," in caps_raw else caps_raw
    else:
        capabilities = None
    user_id = auth_service.resolve_user_from_external(
        issuer="dev",
        subject=subject,
        email=email,
        display_name=display_name,
    )
    return auth_service.mint_token(user_id, role=role, ns=ns, capabilities=capabilities)


@router.get("/login")
async def login(
    request: Request,
    redirect_uri: str | None = None,
    return_to: str | None = None,
    provider: str | None = None,
) -> RedirectResponse:
    auth_mode = _get_auth_mode(request)

    # redirect_uri override is a dev-only convenience — reject it in prod
    if redirect_uri and auth_mode != "dev":
        raise HTTPException(
            status_code=400,
            detail="redirect_uri override is only allowed when AUTH_MODE=dev",
        )

    if auth_mode == "dev":
        auth_service = get_auth_service(request)
        ui_url = redirect_uri or _resolve_return_target(return_to)
        token = _mint_dev_token(auth_service)
        return RedirectResponse(f"{ui_url}?token={token}")

    ui_url = _resolve_return_target(return_to)
    provider_name, oidc = _get_oidc_client_for_provider(request, provider)
    state = OIDCClient.generate_state()
    pkce = OIDCClient.generate_pkce()
    nonce = OIDCClient.generate_nonce()

    auth_url = await oidc.get_auth_url(state, pkce["challenge"], nonce)

    response = RedirectResponse(auth_url)
    response.set_cookie("auth_state", state, httponly=True, max_age=600)
    response.set_cookie(
        "auth_pkce_verifier", pkce["verifier"], httponly=True, max_age=600
    )
    response.set_cookie("auth_nonce", nonce, httponly=True, max_age=600)
    response.set_cookie("auth_provider", provider_name, httponly=True, max_age=600)
    response.set_cookie("auth_return_to", ui_url, httponly=True, max_age=600)
    return response


@router.get("/callback")
async def callback(
    request: Request,
    state: str,
    code: str | None = None,
    error: str | None = None,
    error_description: str | None = None,
) -> RedirectResponse:
    auth_mode = _get_auth_mode(request)
    if auth_mode == "dev":
        raise HTTPException(
            status_code=400, detail="OIDC callback disabled (AUTH_MODE=dev)"
        )

    if error:
        raise HTTPException(
            status_code=400, detail=f"OIDC Error: {error} - {error_description}"
        )

    if not code:
        raise HTTPException(status_code=400, detail="Missing authorization code")

    auth_service = get_auth_service(request)
    stored_state = request.cookies.get("auth_state")
    stored_verifier = request.cookies.get("auth_pkce_verifier")
    stored_nonce = request.cookies.get("auth_nonce")
    stored_provider = request.cookies.get("auth_provider")
    stored_return_to = request.cookies.get("auth_return_to")

    if not stored_state or state != stored_state:
        raise HTTPException(status_code=400, detail="Invalid state")
    if not stored_verifier:
        raise HTTPException(status_code=400, detail="Missing PKCE verifier")
    if not stored_nonce:
        raise HTTPException(status_code=400, detail="Missing nonce")
    if not stored_provider:
        raise HTTPException(status_code=400, detail="Missing provider")
    return_target = _resolve_return_target(stored_return_to)

    _, oidc = _get_oidc_client_for_provider(request, stored_provider)

    tokens = await oidc.exchange_code(code, stored_verifier)
    id_token = tokens.get("id_token")
    if not isinstance(id_token, str) or not id_token:
        raise HTTPException(status_code=400, detail="Missing id_token")

    try:
        id_claims = await oidc.validate_id_token(id_token, nonce=stored_nonce)
    except ValueError as exc:
        raise HTTPException(status_code=401, detail=str(exc)) from exc

    access_token = tokens.get("access_token")
    if not isinstance(access_token, str) or not access_token:
        raise HTTPException(status_code=400, detail="Missing access_token")
    userinfo = await oidc.get_userinfo(access_token)
    subject = _required_string(id_claims, "sub")
    if userinfo.get("sub") != subject:
        raise HTTPException(
            status_code=401,
            detail="userinfo subject does not match validated id_token subject",
        )

    id_email = _optional_string(id_claims, "email")
    userinfo_email = _optional_string(userinfo, "email")
    if id_email and userinfo_email and id_email != userinfo_email:
        raise HTTPException(
            status_code=401,
            detail="userinfo email does not match validated id_token email",
        )
    email = id_email or userinfo_email
    if not email:
        raise HTTPException(status_code=400, detail="OIDC identity missing email")
    display_name = _optional_string(id_claims, "name") or _optional_string(
        userinfo, "name"
    )

    issuer = _optional_string(id_claims, "iss") or oidc.discovery_url

    user_id = auth_service.resolve_user_from_external(
        issuer=issuer,
        subject=subject,
        email=email,
        display_name=display_name,
    )

    app_token = auth_service.mint_token(user_id)

    separator = "&" if "?" in return_target else "?"
    response = RedirectResponse(f"{return_target}{separator}token={app_token}")
    response.delete_cookie("auth_state")
    response.delete_cookie("auth_pkce_verifier")
    response.delete_cookie("auth_nonce")
    response.delete_cookie("auth_provider")
    response.delete_cookie("auth_return_to")
    return response


@router.get("/me")
async def me(
    request: Request, auth_service: AuthService = Depends(get_auth_service)
) -> dict[str, JsonValue]:
    claims = getattr(request.state, "claims", None)
    if not claims or "user_id" not in claims:
        raise HTTPException(status_code=401, detail="Not authenticated")

    user = auth_service.get_user(claims["user_id"])
    if not user:
        raise HTTPException(status_code=404, detail="User not found")

    return {**user, "role": claims.get("role"), "ns": claims.get("ns")}


@router.post("/logout")
async def logout() -> dict[str, bool]:
    return {"ok": True}